KiCAD Files Completed and General UPA Update
KiCAD design files for the base module and UPA module are completed and on order.

The UPA sensor code is now fully operational. One issue: accuracy of the beam forming. A recalculation of delay timing is required. EEPROM does not seem to write/read properly on breadboard, though this could be to faulty connections. Further testing is required.

KNOWN ISSUES:
* UPA beam forming not accurate
* EEPROM fails to read/write properly.